+++ packages/web-components/src/components/cta-block/__stories__/README.stories.mdx
@@ -1,0 +1,62 @@
+import { Props, Description } from '@storybook/addon-docs/blocks';
+import contributing from '../../../../../../docs/contributing-license.md';
+import markdownContents from '../../../../docs/markdown-contents.md';
+
+# `<dds-cta-section>`
+
+> Use the CTASection to provide additional contextual information for the section or page with specific calls to action.
+> 💡 Check our
+> [CodeSandbox](https://githubbox.com/carbon-design-system/carbon-for-ibm-dotcom/tree/master/packages/web-components/examples/codesandbox/components/cta-section)
+> example implementation.
+[![Edit @carbon/ibmdotcom-web-components](https://codesandbox.io/static/img/play-codesandbox.svg)](https://githubbox.com/carbon-design-system/carbon-for-ibm-dotcom/tree/master/packages/web-components/examples/codesandbox/components/cta-section)
+
+## Getting started
+
+Here's a quick example to get you started.
+
+##### JS
+
+```javascript
+
+import '@carbon/ibmdotcom-web-components/es/components/cta-section/cta-section.js';
+import '@carbon/ibmdtocom-web-components/es/components/cta-section/cta-section-copy.js';
+import '@carbon/ibmdtocom-web-components/es/components/cta-section/cta-section-item.js';
+import '@carbon/ibmdtocom-web-components/es/components/cta-section/cta-section-heading.js';
+import '@carbon/ibmdtocom-web-components/es/components/cta-section/cta-section-item-copy.js';
+import '@carbon/ibmdotcom-web-components/es/components/content-block/content-block-heading.js';
+import '@carbon/ibmdotcom-web-components/es/components/button-group/button-group.js';
+import '@carbon/ibmdotcom-web-components/es/components/button-group/button-group-item.js';
+
+```
+
+### HTML
+
+```html
+<dds-cta-section>
+  <dds-content-block-heading>Curabitur malesuada varius mi eu posuere</dds-content-block-heading>
+  <dds-cta-section-copy>Lorem ipsum *dolor* sit amet</dds-cta-section-copy>
+  <dds-button-group slot='buttons'>
+    <dds-button-group-item href="https://example.com">
+      Primary Button
+    </dds-button-group-item>
+    <dds-button-group-item href="https://example.com">
+      Primary button
+    </dds-button-group-item>
+  </dds-button-group>
+</dds-cta-section>
+```
+
+## Props
+
+<Props of="dds-cta-section" />
+<Description markdown={markdownContents} />
+
+## Stable selectors
+
+See [our README](https://github.com/carbon-design-system/carbon-for-ibm-dotcom/tree/master/packages/web-components#stable-selectors-for-analytics-and-integratione2e-testing-in-web-components) to see how Web Components selector and `data-autoid` should be used.
+
+| Web Components selector       | Compatibility selector                    | Description |
+| ----------------------------- | ----------------------------------------- | ----------- |
+| `<dds-cta-section>`           | `data-autoid="dds--cta-section"`          | Component   |
+
+<Description markdown={contributing} />
