+++ packages/web-components/src/components/cta-block/cta-block.scss
@@ -1,0 +1,93 @@
+//
+// Copyright IBM Corp. 2020, 2021
+//
+// This source code is licensed under the Apache-2.0 license found in the
+// LICENSE file in the root directory of this source tree.
+//
+
+@import '@carbon/ibmdotcom-styles/scss/components/cta-block/cta-block';
+@import '@carbon/ibmdotcom-styles/scss/components/layout/layout';
+
+:host(#{$dds-prefix}-cta-block) {
+  padding-left: $carbon--layout-01;
+  @include cta-block--pattern;
+
+  display: block;
+
+  ::slotted([slot='action']) {
+    margin-top: $carbon--layout-01;
+    padding-bottom: $carbon--layout-03;
+
+    @include carbon--breakpoint('md') {
+      padding-bottom: $carbon--layout-05;
+    }
+
+    @include carbon--breakpoint('lg') {
+      padding-bottom: $carbon--layout-06;
+    }
+  }
+
+  ::slotted(#{$dds-prefix}-content-block-copy) {
+    @include carbon--type-style('expressive-heading-03', true);
+
+    display: inline-block;
+    max-width: 640px;
+
+    @include carbon--breakpoint('sm') {
+      width: 100%;
+    }
+
+    @include carbon--breakpoint('md') {
+      width: 90%;
+    }
+  }
+
+  ::slotted(#{$dds-prefix}-content-item) {
+    width: 100%;
+    margin-top: $carbon--spacing-07;
+    margin-bottom: 0;
+    padding-bottom: $carbon--spacing-05;
+    position: relative;
+    display: flex;
+    flex-direction: column;
+    @include carbon--make-col-ready;
+  }
+
+  ::slotted(#{$dds-prefix}-content-block-heading) {
+    @include dds--make-col(7, 12);
+    margin-left: -$carbon--spacing-02;
+  }
+
+  ::slotted(#{$dds-prefix}-content-block-copy) {
+    @include dds--make-col(7, 12);
+  }
+
+  ::slotted(#{$dds-prefix}-text-cta) {
+    margin-left: $carbon--spacing-05;
+  }
+
+  .#{$prefix}--helper-wrapper > .#{$prefix}--content-item-wrapper {
+    display: grid;
+    grid-template-columns: 1fr;
+    grid-template-rows: auto;
+
+    @include carbon--breakpoint('md') {
+      grid-template-columns: 1fr 1fr 1fr;
+    }
+  }
+}
+
+:host(#{$dds-prefix}-cta-section-item-heading) {
+  @include carbon--type-style('expressive-heading-02', true);
+
+  display: block;
+  width: 100%;
+
+  @include carbon--breakpoint('md') {
+    width: 90%;
+  }
+}
+
+.#{$dds-prefix}-ce--cta-block__col {
+  @include dds--make-col(2, 3);
+}
