--- packages/web-components/src/components/cta-section/cta-section-item-copy.ts
@@ -1,38 +1,0 @@
-/**
- * @license
- *
- * Copyright IBM Corp. 2020
- *
- * This source code is licensed under the Apache-2.0 license found in the
- * LICENSE file in the root directory of this source tree.
- */
-
-import { html, customElement } from 'lit-element';
-import ddsSettings from '@carbon/ibmdotcom-utilities/es/utilities/settings/settings.js';
-import DDSContentItemCopy from '../content-item/content-item-copy';
-import StableSelectorMixin from '../../globals/mixins/stable-selector';
-import styles from './cta-section.scss';
-
-const { stablePrefix: ddsPrefix } = ddsSettings;
-
-/**
- * The copy content of CTA section item.
- *
- * @element dds-cta-section-item-copy
- */
-@customElement(`${ddsPrefix}-cta-section-item-copy`)
-class DDSCTASectionItemCopy extends StableSelectorMixin(DDSContentItemCopy) {
-  render() {
-    return html`
-      <slot></slot>
-    `;
-  }
-
-  static get stableSelector() {
-    return `${ddsPrefix}-cta-section-item__copy`;
-  }
-
-  static styles = styles; // `styles` here is a `CSSResult` generated by custom WebPack loader
-}
-
-export default DDSCTASectionItemCopy;
