--- packages/web-components/src/components/cta-section/cta-section-item.ts
@@ -1,43 +1,0 @@
-/**
- * @license
- *
- * Copyright IBM Corp. 2020, 2021
- *
- * This source code is licensed under the Apache-2.0 license found in the
- * LICENSE file in the root directory of this source tree.
- */
-
-import { html, css, customElement } from 'lit-element';
-import ddsSettings from '@carbon/ibmdotcom-utilities/es/utilities/settings/settings.js';
-import StableSelectorMixin from '../../globals/mixins/stable-selector';
-import styles from './cta-section.scss';
-import DDSContentItem from '../content-item/content-item';
-
-const { stablePrefix: ddsPrefix } = ddsSettings;
-
-/**
- * List item version of Content Item.
- *
- * @element dds-cta-section-item
- */
-@customElement(`${ddsPrefix}-cta-section-item`)
-class DDSCTASectionItem extends StableSelectorMixin(DDSContentItem) {
-  render() {
-    return html`
-      <slot name="heading"></slot>
-      <slot></slot>
-      <slot name="footer"></slot>
-    `;
-  }
-
-  static get stableSelector() {
-    return `${ddsPrefix}--cta-section-item`;
-  }
-
-  // `styles` here is a `CSSResult` generated by custom WebPack loader
-  static get styles() {
-    return css`${super.styles}${styles}`;
-  }
-}
-
-export default DDSCTASectionItem;
